Make pcal recognise the start day
The default behaviour of the Unix utility "cal" is to start the week
on the day that corresponds to the LC_TIME locale. Under en_GB, it's
Monday, under en_US, it's Sunday (and under cs_CZ, it's Pondělí, which
corresponds to Monday). Adding the third argument to pcal makes it
possible to emulate at least both the English approaches.

The expected behaviour is the following commands should succeed:

  pcal 7 2023 1 | diff -wB - <(LC_ALL=en_GB.utf8 cal)
  pcal 7 2023 0 | diff -wB - <(LC_ALL=en_US.utf8 cal)